requirements.txt

python-telegram-bot
Django==1.10.5
django-reversion==2.0.8
